very large jobs (eg. 1000 nodes) can make the toolTip way bigger than the browser. this is hard to view, kinda useless to see, and slow to render.
limit the number of nodes displayed in the toolTip by showing only first and last nodes and the averages. in particular, jobs of >128 nodes are displayed as 50 first nodes, 50 last nodes and averages.

could just do this just by setting visibility:hidden on the divs and fudging positions, but 1k nodes worth of divs still takes a while to draw. so instead keep an array of hiddenNodes and don't draw or update any the middle divs. we still need to keep and update the data for the middle nodes in order to get the averages correct, but no longer render them.

the '...' line between the top and the tail is pretty lame at the moment with a very fat a clunky looking border:dashed. needs improvement.