tidy some display code, and also display a whole machine filesystem summary line.
the fs summary line is only a sum across fs stats that the server code has found in jobs.
because most types of sub-node jobs are excluded by the server, this is actually an under-estimate of the total fs activity.
also some care is taken to only include fs stats from each node once to avoid double-counting.

a more accurate total fs summary requires changes to the server code and the xml protocol.
still, this is better than nothing, and makes it look more like 'top'.