allow multiple gmond sources
an arbitrary number of gmond sources can be queried and the data from
all of them is merged together. this can be useful if eg. your storage
cluster has a different ganglia configuration from the compute nodes.
the gmond source for each node is recorded. <gmondGroup> is added to
the protocol so that the ganglia URL for the node can be determined on
the client.