improve bonus node and core accounting
the UserPie graph sits at 100% used when there are a bunch of bonus
cores being used. try to fix this by adding to avail cores/nodes in
a similar way to how we add to used.
ie. if a job is using a core on a "bonus" node, then add that node to
the available count and add all the node's cores to available too.

on server side this is for history graph scaling (I think).
on client side it's for the UserPie.